- Разделил модель на разные файлы для лучшей гибкости кода.
