+++ Tracker/Trackers/TrackersDataStorage/TrackerCategoryModels.swift
@@ -1,0 +1,13 @@
+//
+//  TrackerCategoryModels.swift
+//  Tracker
+//
+//  Created by Вадим on 11.08.2024.
+//
+
+import UIKit
+
+struct TrackerCategory {
+    let title: String
+    let trackers: [Tracker]
+}
