+++ Tracker/Trackers/TrackersDataStorage/TrackerRecordModels.swift
@@ -1,0 +1,13 @@
+//
+//  TrackerRecordModels.swift
+//  Tracker
+//
+//  Created by Вадим on 11.08.2024.
+//
+
+import UIKit
+
+struct TrackerRecord: Hashable {
+    let trackerId: UUID
+    let date: String
+}
